allow browser refresh to get preserved context
When multiple contexts are being preserved, do not purge the context immediately after use since a browser refresh could want that context again.  The oldest context will be purged by getNextPreserveKeyAndPurgeOld once the maximum number preserved has been hit.  This is how this used to behave, but was altered by a fix for when maxNumContextsPreserved eq 1.