Ensure request.context is available in onError handler
IF an error occurs early enough in the application lifecycle there might not be a request.context in the onError handler, this causes an error when the error handler controller method is called.  This change initializes an empty request.context if it does not yet exist in the onError handler.