CF10/Tomcat CGI.PATH_INFO FIX and Route End of String now allowed
In ColdFusion 10 the CGI.PATH_INFO variable always comes back empty.
This simple fix makes sure to use the variable set the URL rewriters.

I wrote about this @
http://www.giancarlogomez.com/2012/06/you-are-not-going-crazy-cgipathinf
o-is.html

Also, routes can now assign a $ to the end of a pattern and it will
respect it and not add the extra "/" or (.*) after it so pattern can
match an end of string match