Allow for end of string "$" regex rule in Routes
When a route gets processed it automatically appends "/(.*)" to it to
pass to the processRoutes method. If you wanted to specify a route like
"^admin/$" wanting to match "/admin/" only it would give you false
positives as it would match more than that due to the appended values.
Now it will respect if you add a "$" to the end of your route.