Add file-based locking around ns install
At World Singles we discovered that if you load cfmljure into multiple
applications running in the same JVM instance, you could get conflicts
while Leiningen is running since it runs file-system destructive
operations. This code uses a lock file in the temporary directory as a
flag to indicate another process is running.

Note that a named lock is *not* sufficient as you can have multiple web
contexts in a single JVM instance (separate CFML server instances).

Downside: it is possible for initialization to fail and leave the lock
file in place (at least in theory) so you may find an application will
not start and you'll have to manually delete the lock file (look in the
console log to see whether it is waiting for the lock file to be deleted).