Changed "folderList" to "folders" in findProjectFile()
It looks like "folderList" is now "folders", which was referenced in the throw(). This results in a variable does not exist error if the throw is triggered.

Changed "folderList" to "folders" and converted it from an array to a list so it can be displayed in the thrown message string.