Fix #333 by supporting .lc / .lucee file extensions
Note that because both .lc and .lucee files can have classes or template
code in them, we have to search for both in all cases. We also have a
restriction that FW/1 (unused) index page _must_ still be `index.cfm`,
because we can't "unhandle" a path that would overlap with the index
page which FW/1 _must_ handle by definition.