Update box.json
After some trial and error with a test package on ForgeBox, I believe this box.json should work properly as far as making sure that only the /framework folder is pulled and it is indeed called "framework". I had some issues with another package having "master" appended to it and then a situation where only the contents of the dir were pulled.