Fix #379 so subsystem delimiter omitted
There was an edge case -- where the requested action included the
delimiter -- where this method could include the delimiter when it was
not supposed to. I believe it now correctly takes both the request and
the implied current action into account!

Also fixes default for home action (which previously did not call
getFullyQualifiedAction() when it should have!).