Fix #394 by actually throwing an exception!
The code was in place to catch and wrap an exception such as this in
order to make it easier to debug. Unfortunately it was missing the
actual throw statement!

I also tidied up the reporting from exceptions that occur when trying to
get a component's metadata: it now attempts to give you the template and
line number where the underlying error occurred.