Proof of concept to show how Clojure can call back into CFML!
The global `before()` function stores the PageContext and the bean
factory into RC so they are available in Clojure controllers.

The `main` controller contains `call-method` which accepts RC, a CFC
instance, a method name, and whatever arguments you want to pass to
CFML, and it uses the PageContext in the RC to make the call.

Then it has `get-bean` which uses `call-method` to call `getBean()` on
the bean factory.