Address #449 by not caching bean in accumulator
This modifies `resolveBeanCreate()` to no longer cache the bean in the
accumulator and instead return a fresh structure each time (whilst still
caching the other metadata from the accumulator).

This should solve the general problem but still leaves a potential edge
case when transients have transients as dependencies (because the
injected transients are still going to be cached... I think).