Remove Gitter badge/link? (#536)
The last activity in the Gitter room was July 2020 (a year ago) and that had been a year since any prior messages. Rather than risk fragment an already fairly small community, perhaps the prominent Gitter badge and link should just be removed from the README.

I'm not sure if it is referenced from anywhere else.